requirements.txt
numpy>=1.26.0
scipy>=1.13.0
matplotlib>=3.9.0
python-dateutil>=2.9.0
rich>=13.7.0
pytest>=8.0.0
python-telegram-bot>=21.0.0
pyyaml>=6.0
colorama>=0.4.6
